refactor: AccountQueryClient
